// androidtv-remote.d.ts
    export class AndroidRemote {
        constructor(...args: any[]);

        getCertificate(...args: any[]): void;

        sendAppLink(...args: any[]): void;

        sendCode(...args: any[]): void;

        sendKey(...args: any[]): void;

        sendPower(...args: any[]): void;

        start(...args: any[]): void;

        stop(...args: any[]): void;

        static captureRejectionSymbol: any;

        static captureRejections: boolean;

        static defaultMaxListeners: number;

        static errorMonitor: any;

        static getEventListeners(emitterOrTarget: any, type: any): any;

        static init(opts: any): void;

        static kMaxEventTargetListeners: any;

        static kMaxEventTargetListenersWarned: any;

        static listenerCount(emitter: any, type: any): any;

        static on(emitter: any, event: any, options: any): any;

        static once(emitter: any, name: any, options: any): any;

        static setMaxListeners(n: any, eventTargets: any): void;

        static usingDomains: boolean;

    }

    export const RemoteDirection: {
        "0": string;
        "1": string;
        "2": string;
        "3": string;
        END_LONG: number;
        SHORT: number;
        START_LONG: number;
        UNKNOWN_DIRECTION: number;
    };

    export const RemoteKeyCode: {
// ...
        KEYCODE_BACK: number;
// ...
        KEYCODE_BUTTON_SELECT: number;
// ...
        KEYCODE_CHANNEL_DOWN: number;
        KEYCODE_CHANNEL_UP: number;
// ...
        KEYCODE_DPAD_CENTER: number;
        KEYCODE_DPAD_DOWN: number;
        KEYCODE_DPAD_DOWN_LEFT: number;
        KEYCODE_DPAD_DOWN_RIGHT: number;
        KEYCODE_DPAD_LEFT: number;
        KEYCODE_DPAD_RIGHT: number;
        KEYCODE_DPAD_UP: number;
// ...
        KEYCODE_HOME: number;
// ...
        KEYCODE_MEDIA_FAST_FORWARD: number;
        KEYCODE_MEDIA_NEXT: number;
        KEYCODE_MEDIA_PAUSE: number;
        KEYCODE_MEDIA_PLAY: number;
        KEYCODE_MEDIA_PLAY_PAUSE: number;
        KEYCODE_MEDIA_PREVIOUS: number;
        KEYCODE_MEDIA_RECORD: number;
        KEYCODE_MEDIA_REWIND: number;
        KEYCODE_MEDIA_SKIP_BACKWARD: number;
        KEYCODE_MEDIA_SKIP_FORWARD: number;
        KEYCODE_MEDIA_STEP_BACKWARD: number;
        KEYCODE_MEDIA_STEP_FORWARD: number;
        KEYCODE_MEDIA_STOP: number;
        KEYCODE_MEDIA_TOP_MENU: number;
        KEYCODE_MENU: number;
        KEYCODE_META_LEFT: number;
        KEYCODE_META_RIGHT: number;
        KEYCODE_MINUS: number;
        KEYCODE_MOVE_END: number;
        KEYCODE_MOVE_HOME: number;
        KEYCODE_MUHENKAN: number;
        KEYCODE_MUSIC: number;
        KEYCODE_MUTE: number;
// ...
        KEYCODE_TV: number;
        KEYCODE_TV_ANTENNA_CABLE: number;
        KEYCODE_TV_AUDIO_DESCRIPTION: number;
        KEYCODE_TV_AUDIO_DESCRIPTION_MIX_DOWN: number;
        KEYCODE_TV_AUDIO_DESCRIPTION_MIX_UP: number;
        KEYCODE_TV_CONTENTS_MENU: number;
        KEYCODE_TV_DATA_SERVICE: number;
        KEYCODE_TV_INPUT: number;
        KEYCODE_TV_INPUT_COMPONENT_1: number;
        KEYCODE_TV_INPUT_COMPONENT_2: number;
        KEYCODE_TV_INPUT_COMPOSITE_1: number;
        KEYCODE_TV_INPUT_COMPOSITE_2: number;
        KEYCODE_TV_INPUT_HDMI_1: number;
        KEYCODE_TV_INPUT_HDMI_2: number;
        KEYCODE_TV_INPUT_HDMI_3: number;
        KEYCODE_TV_INPUT_HDMI_4: number;
        KEYCODE_TV_INPUT_VGA_1: number;
// ...
        KEYCODE_VOLUME_DOWN: number;
        KEYCODE_VOLUME_MUTE: number;
        KEYCODE_VOLUME_UP: number;
        KEYCODE_W: number;
        KEYCODE_WAKEUP: number;
        KEYCODE_WINDOW: number;
        KEYCODE_X: number;
        KEYCODE_Y: number;
        KEYCODE_YEN: number;
        KEYCODE_Z: number;
        KEYCODE_ZENKAKU_HANKAKU: number;
        KEYCODE_ZOOM_IN: number;
        KEYCODE_ZOOM_OUT: number;
    };